Move to hibernate package
Can't have two major versions of Hibernate anyway since they have
the same base org.hibernate package.
