+++ src/test/java/org/kiwiproject/hibernate/CriteriaQueriesTest.java
@@ -1,4 +1,4 @@
-package org.kiwiproject.hibernate5;
+package org.kiwiproject.hibernate;
 
 import static com.google.common.collect.Lists.newArrayList;
 import static org.assertj.core.api.Assertions.assertThat;
